Migrate unit tests that are still relevant after summarymap is removed.
