+++ config-model/src/test/java/com/yahoo/schema/derived/SummaryTestCase.java
@@ -2,19 +2,29 @@
 package com.yahoo.schema.derived;
 
 import com.yahoo.config.model.application.provider.BaseDeployLogger;
+import com.yahoo.config.model.test.MockApplicationPackage;
+import com.yahoo.document.PositionDataType;
+import com.yahoo.schema.RankProfileRegistry;
 import com.yahoo.schema.Schema;
 import com.yahoo.schema.ApplicationBuilder;
 import com.yahoo.schema.AbstractSchemaTestCase;
+import com.yahoo.schema.document.SDDocumentType;
+import com.yahoo.schema.document.SDField;
 import com.yahoo.schema.parser.ParseException;
+import com.yahoo.schema.processing.Processing;
 import com.yahoo.vespa.config.search.SummaryConfig;
+import com.yahoo.vespa.documentmodel.SummaryTransform;
+import com.yahoo.vespa.model.container.search.QueryProfiles;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
-import java.util.Iterator;
+import java.util.Set;
 
 import static com.yahoo.config.model.test.TestUtil.joinLines;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * Tests summary extraction
@@ -60,10 +70,7 @@ public class SummaryTestCase extends AbstractSchemaTestCase {
         SummaryClass summary = new SummaryClass(schema, schema.getSummary("default"), new BaseDeployLogger());
         assertEquals("default", summary.getName());
 
-        Iterator<SummaryClassField> fields = summary.fields().values().iterator();
-
-        SummaryClassField field;
-
+        var fields = summary.fields().values().iterator();
         assertEquals(13, summary.fields().size());
 
         assertSummaryField("exactemento", SummaryClassField.Type.LONGSTRING, fields.next());
@@ -155,6 +162,102 @@ public class SummaryTestCase extends AbstractSchemaTestCase {
         var summary = new SummaryClass(schema, schema.getSummary(summaryName), new BaseDeployLogger());
         var config = new SummaryConfig.Classes(summary.getSummaryClassConfig());
         assertEquals(expected, config.omitsummaryfeatures());
+    }
+
+    @Test
+    void testPositionDeriving() {
+        Schema schema = new Schema("store", MockApplicationPackage.createEmpty());
+        SDDocumentType document = new SDDocumentType("store");
+        schema.addDocument(document);
+        String fieldName = "location";
+        SDField field = document.addField(fieldName, PositionDataType.INSTANCE);
+        field.parseIndexingScript("{ attribute | summary }");
+        new Processing().process(schema, new BaseDeployLogger(), new RankProfileRegistry(), new QueryProfiles(),
+                true, false, Set.of());
+
+        var summary = new SummaryClass(schema, schema.getSummary("default"), new BaseDeployLogger());
+        var fields = summary.fields().values().iterator();
+        assertEquals(4, summary.fields().size());
+        assertSummaryField(fieldName, SummaryClassField.Type.JSONSTRING, "geopos", "location_zcurve", fields.next());
+        assertSummaryField("rankfeatures", SummaryClassField.Type.FEATUREDATA, "rankfeatures", fields.next());
+        assertSummaryField("summaryfeatures", SummaryClassField.Type.FEATUREDATA, "summaryfeatures", fields.next());
+        assertSummaryField("documentid", SummaryClassField.Type.LONGSTRING, "documentid", fields.next());
+    }
+
+    @Test
+    void testFailOnSummaryFieldSourceCollision() {
+        try {
+            ApplicationBuilder.buildFromFile("src/test/examples/summaryfieldcollision.sd");
+        } catch (Exception e) {
+            assertTrue(e.getMessage().matches(".*equally named field.*"));
+        }
+    }
+
+    @Test
+    void source_field_is_passed_as_argument_in_matched_elements_filter_transforms() throws ParseException {
+        assertOverride(joinLines("field my_field type map<string, string> {",
+                "  indexing: summary",
+                "  summary: matched-elements-only",
+                "  struct-field key { indexing: attribute }",
+                "}"), "my_field", SummaryTransform.MATCHED_ELEMENTS_FILTER.getName());
+
+        assertOverride(joinLines("field my_field type map<string, string> {",
+                "  indexing: summary",
+                "  summary: matched-elements-only",
+                "  struct-field key { indexing: attribute }",
+                "  struct-field value { indexing: attribute }",
+                "}"), "my_field", SummaryTransform.MATCHED_ATTRIBUTE_ELEMENTS_FILTER.getName());
+    }
+
+    @Test
+    void commands_that_are_dynamic_and_require_the_query() {
+        assertTrue(SummaryMap.isDynamicCommand("dynamicteaser"));
+        assertTrue(SummaryMap.isDynamicCommand(SummaryTransform.MATCHED_ELEMENTS_FILTER.getName()));
+        assertTrue(SummaryMap.isDynamicCommand(SummaryTransform.MATCHED_ATTRIBUTE_ELEMENTS_FILTER.getName()));
+        assertFalse(SummaryMap.isDynamicCommand(SummaryTransform.ATTRIBUTE.getName()));
+    }
+
+    @Test
+    void documentid_summary_field_has_corresponding_summary_transform() throws ParseException {
+        var schema = buildSchema("field foo type string { indexing: summary }",
+                joinLines("document-summary bar {",
+                        "    summary documentid type string {}",
+                        "}"));
+        assertOverride(schema, "documentid", SummaryTransform.DOCUMENT_ID.getName(), "", "bar");
+    }
+
+    @Test
+    void documentid_summary_transform_requires_disk_access() {
+        assertFalse(SummaryTransform.DOCUMENT_ID.isInMemory());
+    }
+
+    private void assertOverride(String fieldContent, String expFieldName, String expCommand) throws ParseException {
+        assertOverride(buildSchema(fieldContent, ""), expFieldName, expCommand, expFieldName);
+    }
+
+    private void assertOverride(Schema schema, String expFieldName, String expCommand, String expSource) throws ParseException {
+        assertOverride(schema, expFieldName, expCommand, expSource, "default");
+    }
+
+    private void assertOverride(Schema schema, String expFieldName, String expCommand, String expSource, String summaryClass) throws ParseException {
+        var summary = new SummaryClass(schema, schema.getSummary(summaryClass), new BaseDeployLogger());
+        var cfg = new SummaryConfig.Classes(summary.getSummaryClassConfig());
+        var field = cfg.fields(0);
+        assertEquals(expFieldName, field.name());
+        assertEquals(expCommand, field.command());
+        assertEquals(expSource, field.source());
+    }
+
+    private Schema buildSchema(String field, String documentSummary) throws ParseException {
+        var builder = new ApplicationBuilder(new RankProfileRegistry());
+        builder.addSchema(joinLines("search test {",
+                "  document test {",
+                field,
+                "  }",
+                documentSummary,
+                "}"));
+        builder.build(true);
+        return builder.getSchema();
     }
 
 }
